feat: initial support for recursive product types
This commit implements initial support for recursive product data types.
In C, they're represented as structs that have a field that is a pointer
to the same struct type.

In Carp, we currently substitute recursive references with pointers to
the type, and users must provide a pointer argument during
instantiation. To make creating initial values of these types easier, we
define a make function, which initializes a value of the type with its
recursive part set to the null pointer.