fix: don't hang on module expansion
Our macro/symbol expansion in Expand.hs was missing a module case. This
would cause hangs whenever a user erroneously passed a module name as an
argument to a function, since expand just returns the symbol it received
when the symbol has no known expansion. For example, this code would
cause a hang since `B` is a module:

```clojure
(load "IO.carp")

(register-type A "")

(register-type B [a A])
(defn main []
  (do
    (IO.println "send help")
    (IO.println &(str (width B)))))
```

Instead, we now report a type error (since B is a module, not a
pointer).