fix: make type registration with empty override an error
Since C99, typedefs no longer allow omission of a type specifier
(previously an omission would assume the type was int). In Carp,
registering a type with a blank string override (register-type A "")
produces a typedef with no specifier, which is an error in C after C99.

Since this code will often produce erroneous C and is of little
convenience, this commit makes such a registration an error in the eyes
of the Carp compiler. Users that wish to register a type that's
implicitly declared as int and is only expected to be compatible with
variants of C prior to 99 can still define the type externally and
register it using the opaque registration form `(register-type A)`.