fix: don't generate getters for dummy fields
For empty structs, we generate a dummy field for ANSI C compatibility.
This field needs to be included in initializers for the struct, but
should not be emitted in any other functions. I erroneously included it
in other functions in a previous merge. This commit fixes the issue by
ensuring the dummy field is only included in the struct initializer.