refactor: add resolvers to control lookup order
Adds a new abstraction, the *Resolver* which controls the lookup order
when resolving symbols. Env, provides mechanisms for storing and
retrieving bindings, Context bundles environments together (and provides
some retrieval utilities) and Resolver determines the order in which
multiple lookups should be executed.

The Resolver abstractions allows us to remove much of the gnarly lookup
code that was previously directly defined in the evaluator. The
abstraction also provides a Semigroup instance for the flexible
combination of resolvers, making it significantly easier to try out
different orders in the evaluator.

This commit changes no behavior -- it just uses Resolvers to emulate the
previous lookup order in the evaluator. Future commits will leverage the
abstraction to simplify lookups and refine behaviors where needed.