feat: don't manage blittable types
The `blit` interface, when implemented, indicates that a type is freely
copyable and does not require deletion. It is similar to Rust's Copy
trait and indicates that a type should have copy instead of "move"
semantics. This commit updates the memory management system to ignore
such types, since we do not need to worry about deleting them.

Thus, when using a blittable type, users can freely use a variable of
the type in more than one place.

This commit also fixes a small typo and adds a TODO to warn users when
delete and blit are both implemented (as typically one shouldn't need
delete if a type is truly blittable).