test: check blittable memory behavior in memory.carp
This commit updates memory.carp to test that types that implement blit
are freely copyable (have copy instead of move semantics, are not borrow
checked) and do not cause leaks.

Of course, users can still cause leaks if they define blit on a type
that isn't actually freely copyable (e.g. it owns some data that is heap
allocated), but this test checks the basic behavior.