feat: add fputc wrapper to IO (#1417)
* feat: add fputc wrapper to IO

fputc is a low-level function that writes a single C character to a
file. We can use this as a basis for more elegant APIs.

The signature matches that of the C standard library function for
compatibility, which takes an Int instead of a char. However, the
documentation notes that the int argument is converted to an unsigned
char.

* doc: fix typos for fgetc and fputc docs