docs: update memory.md
I've added some additional context to the doc in efforts to explain the
memory management system to users. After this, I'm going to write a bit
more about the actual implementation of the system as it stands
currently. One of the goals of this effort is to make it easier to
upgrade given an exhaustive account of its present workings.